[AN] refactor: Center Scroll Listener 파일 분리  #359
